+++ main.py
@@ -1,45 +1,24 @@
-import os
-import base64
-from dotenv import load_dotenv
-load_dotenv()
 import requests
+import os
 from voice import Voice
+import elevenlabs as el
+from dialogue import Dialogue
 
-BASE_URL = "https://api.elevenlabs.io"
-headers = {
-    "xi-api-key": os.getenv('ELEVEN_LABS_API_KEY')
-}
-
-# Takes an error response and generates a readable exception
-def GenerateException(response):
-    return Exception(f'Failed with status {response.status_code}: '
-        f'{response.json()['detail']['message']}')
-
-# Returns all cloned voices
 def GetClonedVoices():
-    response = requests.get(BASE_URL + "/v2/voices?category=cloned", 
-        headers=headers)
+    '''
+    Fetches and returns a user's cloned voices
+    Args: N/A
+    Returns:
+        [Voice]: Array of Voice objects
+    '''
+
+    response = requests.get(el.BASE_URL + "/v2/voices?category=cloned", 
+        headers=el.HEADERS)
     response.raise_for_status() # Throws error if status 4xx or 5xx
     voices = []
     for v in response.json()['voices']:
         voices.append(Voice(v))
     return voices
-
-def CreateSpeech(voice_id, text):
-    response = requests.post(BASE_URL + f"/v1/text-to-speech/{voice_id}/with-timestamps",
-        headers=headers,
-        json={
-            "text": f"{text}"
-        })
-    if (response.status_code != 200):
-        raise GenerateException(response)
-    return response
-
-
-def ConvertToMp3(base_64_string, output_path='./audio/output.mp3'):
-    mp3_data = base64.b64decode(base_64_string)
-    with open(output_path, "wb") as mp3_file:
-        mp3_file.write(mp3_data)
 
 def Main():
     if (not os.getenv('ELEVEN_LABS_API_KEY')):
@@ -47,12 +26,18 @@ def Main():
     try:
         voices = GetClonedVoices()
         if (len(voices)):
-            print(voices[0].name)
-            #response = CreateSpeech(voices[1]['voice_id'], 'Hello my fellow YN')
-            #ConvertToMp3(response.json()['audio_base64'])
+            voice1 = voices[0]
+            voice2 = voices[1]
+            d = Dialogue(voice1, voice2, "Yo whats up mr white wanna cook bitch?/Jesse go fucking kill yourself retard")
+            d.CreateAllSpeech()
 
     except requests.exceptions.HTTPError as e:
-        return print(f'Failed with status {e.response.status_code}: '
-        f'{e.response.json()['detail']['message']}')
+        # Failed with status [status]: [message]
+        return print(f'Failed with status {e.response.status_code}:'
+        f' {e.response.json()['detail']['message']}')
+    except NameError as e:
+        print(e)
+    except OSError:
+        print('A system error has occured')
 
 Main()